pyproject.toml
[project]
name = "giulia-ai"
version = "0.1.0"
description = "Elite Executive AI Assistant 'Giulia' based on Gemini 3"
readme = "README.md"
authors = [
    {name = "Raptile Bytez", email = "raptile.bytez@gmail.com"}
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
    "python-dotenv>=1.2.1",
]
